n8n-workflows/requirements-n8n.txt
python-dotenv>=1.0.0
playwright>=1.40.0
